change all database id to long datatype

+++ app/src/main/java/com/example/myapp2/model/dao/QuestDao.kt
@@ -22,5 +22,5 @@ interface QuestDao {
 
     @Transaction
     @Query("SELECT * FROM quest WHERE id = :questId")
-    suspend fun getQuestWithOptions(questId: Int): QuestWithOptions
+    suspend fun getQuestWithOptions(questId: Long): QuestWithOptions
 }
