+++ app/src/main/java/com/example/myapp2/model/dao/QuizDao.kt
@@ -21,5 +21,5 @@ interface QuizDao {
 
     @Transaction
     @Query("SELECT * FROM quiz WHERE id = :quizId")
-    suspend fun getQuizWithQuests(quizId: Int): QuizWithQuests
+    suspend fun getQuizWithQuests(quizId: Long): QuizWithQuests
 }
